Añadido a MainScreen el searchBar. Sin implementar busqueda con API

+++ app/src/main/java/curso/alberto/sunriseweatherapp/data-retrofit/Previccion_Hora.kt
@@ -111,7 +111,7 @@ data class TotalLiquid (
     @SerializedName("UnitType" ) var UnitType: Int? = null
 )
 data class Rain (
-    @SerializedName("Value") var Value: Int? = null,
+    @SerializedName("Value") var Value: Double? = null,
     @SerializedName("Unit" ) var Unit: String? = null,
     @SerializedName("UnitType" ) var UnitType: Int? = null
 )
